fix(pre-checks): resolve file size limit and build validation failures
- Split ArchiveSearchView.tsx into ArchiveResultTables.tsx to stay under 200 lines
- Add missing escalation-related fields to BedGridData and BedWithElapsedTime types as optional to fix tsc errors
- Verified build and tsc --noEmit locally
