+++ src/features/bed-dashboard/types/bed.ts
@@ -56,6 +56,7 @@ export interface Bed {
 export interface BedWithElapsedTime extends Bed {
   elapsedTimeMs: number | null
   isDelayed: boolean
+  isEscalated?: boolean             // flagged when elapsed time exceeds escalation threshold
   // US-1.6: Disposition bottleneck fields
   isDispositionBottleneck: boolean
   dispositionElapsedMs: number | null  // time spent in Decision Made stage specifically
@@ -79,6 +80,8 @@ export interface BedGridData {
   beds: BedWithElapsedTime[]
   stages: Stage[]
   delayThresholdMs: number
+  escalationThresholdMs?: number    // threshold ms before a patient is flagged as escalated
+  escalationCount?: number          // count of currently escalated beds
   bottleneckCount: number  // US-1.6: count of active disposition bottlenecks
 }
